python_modules/automation/requirements.txt
boto3==1.9.47
click>=6.7
faker>=1.0.4
google-cloud-storage
pytablereader
random_useragent>=1.0
requests